fix: shared action bar
